+++ src/QuartoNotebookWorker/src/notebook_metadata.jl
@@ -6,7 +6,11 @@ frontmatter, or implicitly set by Quarto itself at runtime. Note that this is a
 copy of the options used internally and mutation will not affect any other parts
 of notebook evaluation.
 """
-notebook_options() = deepcopy(NotebookState.OPTIONS[])
+function notebook_options()
+    ctx = NotebookState.current_context()
+    ctx === nothing && return Dict{String,Any}()
+    return deepcopy(ctx.options)
+end
 
 """
     cell_options() -> Dict{String,Any}
@@ -15,4 +19,4 @@ The options for the current cell being evaluated. This is a copy of the options
 used internally and mutation will not affect any other parts of notebook
 evaluation.
 """
-cell_options() = deepcopy(NotebookState.CELL_OPTIONS[])
+cell_options() = deepcopy(NotebookState.current_cell_options())